requirements.txt
beautifulsoup4==4.13.3
numpy==2.2.4
pandas==2.2.3
openpyxl
python-dotenv
requests
